pyproject.toml
[project]
name = "production-network-testing-lab"
version = "0.1.0"
description = "Production-grade multi-vendor network testing lab"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "PyYAML>=6.0",
    "hypothesis>=6.92.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.3.0",
    "pytest-timeout>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.3.0",
    "pytest-timeout>=2.1.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "yamllint>=1.35.0",
    "ansible-lint>=24.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]

[tool.coverage.run]
source = ["scripts", "ansible/plugins", "ansible/filter_plugins"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.hypothesis]
max_examples = 100
deadline = 5000

[tool.hypothesis.profiles.ci]
max_examples = 1000
deadline = 10000

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "S",   # flake8-bandit (security)
]

ignore = [
    "E501",  # line too long (handled by formatter)
    "S101",  # use of assert (common in tests)
    "S110",  # try-except-pass (intentional in some scripts)
    "S603",  # subprocess without shell=True (intentional)
    "S607",  # starting process with partial path (intentional)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",  # assert allowed in tests
    "S104",  # binding to 0.0.0.0 in test data
    "S108",  # /tmp usage in tests
    "S110",  # try-except-pass in tests
    "S602",  # shell=True in test subprocess calls
    "ARG",   # unused arguments common in fixtures
    "E402",  # import order in test setup
]
"validation/**/*.py" = [
    "S101",  # assert allowed in tests
    "ARG",   # unused arguments in test helpers
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # assert_used - common in tests
